netfilter: nf_tables: limit maximum number of jumps/gotos per netns
Add a new sysctl:

   net.netfilter.nf_tables_jump_max_netns

which is 256000 by default, because old iptables-nft rulesets are more
likely to have more jumps/gotos compared to native nftables rulesets.

This limit prevents soft lockups on the packet caused by crafted
rulesets with too many jumps.

This jump/goto limit is global for all tables that are defined in the
netns, regardless the family type, but IPv4 and IPv6 are mutually
excluded in the jump counting.

Note that verdict maps count as a single jump/goto because a map lookup
provides a single exact match, therefore, this is equivalent to an
immediate jump to the element (chain) with the highest jump count.

This limit is not the net count of jumps in your ruleset, but the number
of jumps can be visited traversing the acyclic directed graph with
depth-first search. This is done from control plane where the evaluation
of the selectors is not possible, therefore, this represents the
hypothetical worst case.

This patch adds a jump_count[2] field per table which stores the current
number of jumps/gotos in the present [0] and the future [1]. During the
preparation phase, jump_count[1] is updated to store the number of jumps
after the last table validation while processing the batch.

If this batch does not update the number of jumps in this table, then
jump_count[0] provides the current number of jumps and jump_count[1] is
set to -1. When checking if the number of jumps go over the limit, check
if jump_count[1] is >= 0, meaning the number of jumps for this table has
been modified by this batch, otherwise use jump_count[0], meaning this
table has not been modified in terms of new jumps.

After the commit phase, jump_count[0] is set to jump_count[1] if it is
>= 0. Otherwise, in case of abort, jump_count[1] is reset to -1 to
prepare for handling the next batch.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>