netfilter: nf_tables: add missing reset of table valiation state
On abort we should reset the table validation state.
If a transaction fails the final validation early in the commit hook,
the table validation state is changed to NFT_VALIDATE_DO and a replay
of the transaction is triggered: every rule insert will do a graph
validation.  This is much slower, but provides better error reporting
to the user: we can pinpoint the first rule add that introduces a
validation issue.

At this time, we failed to reset the state back to NFT_VALIDATE_SKIP,
i.e., next transaction was in slow-mode from the start.

This is harmless, but it causes longer load times on the first new
transaction after such an error occured.

Signed-off-by: Florian Westphal <fw@strlen.de>